// src/neynarAPI/neynarV2API/openapi/models/signer.ts
/* tslint:disable */
/* eslint-disable */
/**
 * v2 Farcaster
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface Signer
 */
export interface Signer {
  /**
   * UUID of the signer
   * @type {string}
   * @memberof Signer
   */
  signer_uuid: string;
  /**
   *
   * @type {string}
   * @memberof Signer
   */
  public_key: string;
  /**
   *
   * @type {string}
   * @memberof Signer
   */
  status: SignerStatusEnum;
  /**
   *
   * @type {string}
   * @memberof Signer
   */
  signer_approval_url?: string;
  /**
   *
   * @type {string}
   * @memberof Signer
   */
  fid?: string;
}

/**
 * @export
 * @enum {string}
 */
export enum SignerStatusEnum {
  Generated = "generated",
  PendingApproval = "pending_approval",
  Approved = "approved",
  Revoked = "revoked",
}