// src/neynarAPI/neynarV2API/openapi/models/delete-cast-req-body.ts
/* tslint:disable */
/* eslint-disable */
/**
 * v2 Farcaster
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface DeleteCastReqBody
 */
export interface DeleteCastReqBody {
  /**
   * UUID of the signer
   * @type {string}
   * @memberof DeleteCastReqBody
   */
  signer_uuid: string;
  /**
   * Ethereum address
   * @type {string}
   * @memberof DeleteCastReqBody
   */
  target_hash: string;
}